+++ src/replicon_core/dont_replicate.rs
@@ -1,19 +1,18 @@
-use core::panic;
 use std::{any, marker::PhantomData};
 
 use bevy::{ecs::system::EntityCommands, prelude::*};
-
-use super::replication_rules::Replication;
 
 pub trait CommandDontReplicateExt {
     /**
     Disables replication for component `T`.
 
-    May only be called on an entity if  [`Replication`] was inserted on it this tick.
+    May only be called on an entity without `T` or if `T` was inserted on it this tick.
+    May only be called once per entity.
 
     # Panics
 
-    Panics if called on an entity without [`Replication`] or if [`Replication`] was inserted in a different tick.
+    Panics if `debug_assertions` is enabled and if called on this entity once before
+    or if `T` component was present before this tick.
 
     # Examples
 
@@ -47,23 +46,25 @@ pub trait EntityDontReplicateExt {
 
 impl EntityDontReplicateExt for EntityWorldMut<'_> {
     fn dont_replicate<T: Component>(&mut self) -> &mut Self {
-        // SAFETY: world is not mutated and used only to obtain the tick without atomic synchronization.
-        let tick = unsafe { self.world_mut().change_tick() };
+        if cfg!(debug_assertions) {
+            let component_name = any::type_name::<T>();
+            assert!(
+                !self.contains::<DontReplicate<T>>(),
+                "`dont_replicate::<{component_name}>` shouldn't be called twice for the same entity"
+            );
+
+            // SAFETY: world is not mutated and used only to obtain the tick without atomic synchronization.
+            let tick = unsafe { self.world_mut().change_tick() };
+            if let Some(component_ticks) = self.get_change_ticks::<T>() {
+                assert_eq!(
+                    tick,
+                    component_ticks.added_tick(),
+                    "disabling replication for `{component_name}` should be done only with its insertion",
+                );
+            }
+        }
 
         self.insert(DontReplicate::<T>(PhantomData));
-
-        let component_name = any::type_name::<T>();
-        let replication_name = any::type_name::<Replication>();
-        let replication_ticks = self.get_change_ticks::<Replication>().unwrap_or_else(|| {
-            panic!("disabling replication for `{component_name}` should only be done for entities with `{replication_name}`")
-        });
-
-        assert_eq!(
-            tick,
-            replication_ticks.added_tick(),
-            "disabling replication for `{component_name}` should be done only with `{replication_name}` insertion",
-        );
-
         self
     }
 }
@@ -80,12 +81,16 @@ mod tests {
 
     #[test]
     #[should_panic]
-    fn without_replication() {
+    fn called_twice() {
         let mut world = World::new();
 
         let mut queue = CommandQueue::default();
         let mut commands = Commands::new(&mut queue, &world);
-        commands.spawn_empty().dont_replicate::<Transform>();
+        commands
+            .spawn(Transform::default())
+            .dont_replicate::<Transform>()
+            .dont_replicate::<Transform>();
+
         queue.apply(&mut world);
     }
 
@@ -96,7 +101,7 @@ mod tests {
 
         let mut queue = CommandQueue::default();
         let mut commands = Commands::new(&mut queue, &world);
-        let entity = commands.spawn(Replication).id();
+        let entity = commands.spawn(Transform::default()).id();
         queue.apply(&mut world);
 
         world.increment_change_tick();